cast sizeof() with (int) to avoid a warning under MacOS X
git-svn-id: svn://anonscm.debian.org/svn/pcsclite/trunk/PCSC@281 0ce88b0d-b2fd-0310-8134-9614164e65ea

+++ src/utils/installifd.c
@@ -38,7 +38,7 @@ int main(int argc, char *argv[])
 	FILE *fd;
 
 	printf("Please enter a friendly name for your reader (%d char max)\n",
-		sizeof(lpcReader));
+		(int)sizeof(lpcReader));
 	printf("-----> ");
 
 	fgets(lpcReader, sizeof(lpcReader), stdin);
@@ -49,7 +49,7 @@ int main(int argc, char *argv[])
   retrylib:
 
 	printf("Please enter the full path of the readers driver (%d char max)\n",
-		sizeof(lpcLibrary));
+		(int)sizeof(lpcLibrary));
 	printf("Example: %s/librdr_generic.so\n", PCSCLITE_HP_DROPDIR);
 	printf("-----> ");
